requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.10.0
ccxt>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
